# api/openapi.yaml
openapi: 3.0.3
info:
  title: DonoEngine
  description: Go-based comic catalog backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api/v0
paths:
...
  /types/categories:
    get:
      tags:
        - Type
      summary: List category type.
      operationId: listCategoryType
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Category type list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of category type with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of category type with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Type
      summary: Add category type.
      operationId: addCategoryType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewGenericType'
        required: true
      responses:
        '201':
          description: Category type added.
          headers:
            Location:
              description: The path of new category type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /types/categories/{code}:
    get:
      tags:
        - Type
      summary: Get category type.
      operationId: getCategoryType
      parameters:
        - name: code
          in: path
          description: Code of category type to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category type gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Type
      summary: Update category type.
      operationId: updateCategoryType
      parameters:
        - name: code
          in: path
          description: Code of category type to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetGenericType'
        required: true
      responses:
        '200':
          description: Category type updated.
          headers:
            Location:
              description: The path of updated category type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        '204':
          description: Category type unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Type
      summary: Delete category type.
      operationId: deleteCategoryType
      parameters:
        - name: code
          in: path
          description: Code of category type to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category type deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /types/tags:
    get:
      tags:
        - Type
      summary: List tag type.
      operationId: listTagType
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Tag type list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of tag type with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of tag type with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Type
      summary: Add tag type.
      operationId: addTagType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewGenericType'
        required: true
      responses:
        '201':
          description: Tag type added.
          headers:
            Location:
              description: The path of new tag type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /types/tags/{code}:
    get:
      tags:
        - Type
      summary: Get tag type.
      operationId: getTagType
      parameters:
        - name: code
          in: path
          description: Code of tag type to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tag type gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Type
      summary: Update tag type.
      operationId: updateTagType
      parameters:
        - name: code
          in: path
          description: Code of tag type to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetGenericType'
        required: true
      responses:
        '200':
          description: Tag type updated.
          headers:
            Location:
              description: The path of updated tag type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        '204':
          description: Tag type unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Type
      summary: Delete tag type.
      operationId: deleteTagType
      parameters:
        - name: code
          in: path
          description: Code of tag type to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tag type deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /types/comic-relations:
    get:
      tags:
        - Type
      summary: List comic relation type.
      operationId: listComicRelationType
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Comic relation type list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of comic relation type with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of comic relation type with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Type
      summary: Add comic relation type.
      operationId: addComicRelationType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewGenericType'
        required: true
      responses:
        '201':
          description: Comic relation type added.
          headers:
            Location:
              description: The path of new comic relation type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /types/comic-relations/{code}:
    get:
      tags:
        - Type
      summary: Get comic relation type.
      operationId: getComicRelationType
      parameters:
        - name: code
          in: path
          description: Code of comic relation type to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comic relation type gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Type
      summary: Update comic relation type.
      operationId: updateComicRelationType
      parameters:
        - name: code
          in: path
          description: Code of comic relation type to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGenericType'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetGenericType'
        required: true
      responses:
        '200':
          description: Comic relation type updated.
          headers:
            Location:
              description: The path of updated comic relation type.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericType'
        '204':
          description: Comic relation type unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Type
      summary: Delete comic relation type.
      operationId: deleteComicRelationType
      parameters:
        - name: code
          in: path
          description: Code of comic relation type to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comic relation type deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
components:
  schemas:
    Object:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-go-type: uint
          x-go-name: ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - createdAt
...
    GenericType:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            code:
              type: string
            name:
              type: string
          required: 
            - code
            - name
    NewGenericType:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
      required: 
        - code
        - name
    SetGenericType:
      type: object
      properties:
        code:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties: 
            status:
              type: string
            message:
              type: string
          required:
            - status
            - message
      required:
        - error
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT